CODERUB: Separate Logic Test For RetrieveTeacherById Operation - Teachers
