+++ OtripleS.Web.Api.Tests.Unit/Services/Foundations/Teachers/TeacherServiceTests.Logic.RetrieveTeacherById.cs
@@ -1,0 +1,50 @@
+﻿// ---------------------------------------------------------------
+// Copyright (c) Coalition of the Good-Hearted Engineers
+// FREE TO USE AS LONG AS SOFTWARE FUNDS ARE DONATED TO THE POOR
+// ---------------------------------------------------------------
+
+using FluentAssertions;
+using Moq;
+using OtripleS.Web.Api.Models.Teachers;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Xunit;
+
+namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Teachers
+{
+    public partial class TeacherServiceTests
+    {
+        [Fact]
+        public async Task ShouldRetrieveTeacherByIdAsync()
+        {
+            // given
+            DateTimeOffset dateTime = GetRandomDateTime();
+            Teacher randomTeacher = CreateRandomTeacher(dateTime);
+            Guid inputTeacherId = randomTeacher.Id;
+            Teacher storageTeacher = randomTeacher;
+            Teacher expectedTeacher = randomTeacher;
+
+            this.storageBrokerMock.Setup(broker =>
+                broker.SelectTeacherByIdAsync(inputTeacherId))
+                    .ReturnsAsync(storageTeacher);
+
+            // when
+            Teacher actualTeacher =
+                await this.teacherService.RetrieveTeacherByIdAsync(inputTeacherId);
+
+            // then
+            actualTeacher.Should().BeEquivalentTo(expectedTeacher);
+
+            this.storageBrokerMock.Verify(broker =>
+                broker.SelectTeacherByIdAsync(inputTeacherId),
+                    Times.Once);
+
+            this.storageBrokerMock.VerifyNoOtherCalls();
+            this.loggingBrokerMock.VerifyNoOtherCalls();
+            this.dateTimeBrokerMock.VerifyNoOtherCalls();
+        }
+    }
+}
